refactor: split F1 acceptance scenarios
