+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+python_files = test_*.py f[0-9]s*.py
+python_functions = test_* f[0-9]s*
